requirements.txt
django==4.2
djangorestframework==3.14.0
opencv-python==4.8.0.76
numpy==1.24.4
python-decouple==3.8
django-cors-headers==4.2.0